+++ gateway/modules/functions/operations.go
@@ -2,6 +2,9 @@ package functions
 
 import (
 	"context"
+	"fmt"
+
+	"github.com/spaceuptech/helpers"
 
 	"github.com/spaceuptech/space-cloud/gateway/config"
 	"github.com/spaceuptech/space-cloud/gateway/model"
@@ -25,4 +28,17 @@ func (m *Module) AddInternalRule(service string, rule *config.Service) {
 	defer m.lock.Unlock()
 
 	m.config.InternalServices[service] = rule
+}
+
+// GetEndpointContextTimeout returns the endpoint timeout of particular remote-service
+func (m *Module) GetEndpointContextTimeout(ctx context.Context, service, function string) (int, error) {
+	m.lock.RLock()
+	defer m.lock.RUnlock()
+
+	if serviceVal, ok := m.config.Services[service]; ok {
+		if endpoint, ok := serviceVal.Endpoints[function]; ok {
+			return endpoint.Timeout, nil
+		}
+	}
+	return 0, helpers.Logger.LogError(helpers.GetRequestID(ctx), fmt.Sprintf("Could not find endpoint (%s) for service (%s)", function, service), nil, nil)
 }
